requirements.txt

GitPython==3.1.40
PyGithub==1.59.1